+++ CHANGELOG.md
@@ -108,6 +108,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1448 General proto api for NNS libraries 
 -   \#1480 Add return code for AVX512 selection
 -   \#1524 Update config "preload_table" description
+-   \#1537 Optimize raw vector and uids read/write
 -   \#1544 Update resources name in HTTP module
 -   \#1567 Update yaml config description